+++ lib/model/database_handler.dart
@@ -7,10 +7,10 @@ class DatabaseHandler {
   Future<Database> initializeDB() async {
     String path = await getDatabasesPath();
     return openDatabase(
-      join(path, 'userfive.db'),
+      join(path, 'userfiveone.db'),
       onCreate: (database, version) async {
         await database.execute(
-          "CREATE TABLE userfive(id INTEGER PRIMARY KEY AUTOINCREMENT, name TEXT NOT NULL)",
+          "CREATE TABLE userfiveone(id INTEGER PRIMARY KEY AUTOINCREMENT, name TEXT NOT NULL)",
         );
       },
       version: 1,
@@ -21,14 +21,15 @@ class DatabaseHandler {
     int result = 0;
     final Database db = await initializeDB();
     for (var user in users) {
-      result = await db.insert('userfive', user.toMap());
+      result = await db.insert('userfiveone', user.toMap());
     }
     return result;
   }
 
   Future<List<User>> retrieveUsers() async {
     final Database db = await initializeDB();
-    final List<Map<String, Object?>> queryResult = await db.query('userfive');
+    final List<Map<String, Object?>> queryResult =
+        await db.query('userfiveone');
     return queryResult.map((e) => User.fromMap(e)).toList();
   }
 }
