--- lib/view/my_app.dart
@@ -1,33 +1,0 @@
-import 'package:flutter/material.dart';
-import 'expense_first_page.dart';
-
-class ExpenseApp extends StatelessWidget {
-  const ExpenseApp({Key? key}) : super(key: key);
-
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      debugShowCheckedModeBanner: false,
-      theme: ThemeData(
-        primarySwatch: Colors.red,
-      ),
-      home: const ExpensePage(),
-    );
-  }
-}
-
-class ExpensePage extends StatelessWidget {
-  const ExpensePage({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Expense List'),
-      ),
-      body: const UserFirstPage(),
-    );
-  }
-}
